+++ docs/source/changelog.md
@@ -1,5 +1,8 @@
 # Changelog
 
+## Latest
+- Implement support for loading polars dataframes from DuckDB.
+
 ## 0.6.5 (2023-08-16)
 - Implemented automatic versioning of tasks by setting task version to [](#AUTO_VERSION).
   This feature is currently only supported by Polars [`LazyFrame`](inv:pl#reference/lazyframe/index) and by Pandas.